requirements.txt
# Core dependencies
streamlit>=1.31.0
supabase>=2.0.0
python-dotenv>=1.0.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Data visualization
altair>=5.0.0
plotly>=5.13.0
openpyxl>=3.1.0

# Web requests
requests>=2.31.0
uuid>=1.30.0

# UI enhancements
streamlit-extras>=0.3.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Code quality
black>=23.9.0
pylint>=2.17.0

# Utilities
pydantic>=2.0.0
